[#162] refactor: 리졸버를 애플리케이션 단에 위치하도록 리팩터링
