+++ src/test/java/com/kiero/schedule/service/ScheduleServiceTest.java
@@ -56,7 +56,7 @@ import com.kiero.schedule.domain.enums.DayOfWeek;
 import com.kiero.schedule.domain.enums.ScheduleColor;
 import com.kiero.schedule.domain.enums.ScheduleStatus;
 import com.kiero.schedule.domain.enums.StoneType;
-import com.kiero.schedule.domain.enums.TodayScheduleStatus;
+import com.kiero.schedule.application.service.resolver.TodayScheduleStatus;
 
 @ExtendWith(MockitoExtension.class)
 public class ScheduleServiceTest {
